requirements.txt
allure_python_commons
pytest
pytest-html-reporter
allure-pytest
python-dotenv
requests